revert changes to nlp path and add custom path in diffengine-cone-prog
