+++ cvxpy/reductions/solvers/nlp_solvers/diff_engine/__init__.py
@@ -22,7 +22,7 @@ for automatic differentiation.
 """
 
 from cvxpy.reductions.solvers.nlp_solvers.diff_engine.c_problem import C_problem
-from cvxpy.reductions.solvers.nlp_solvers.diff_engine.converters import build_capsule, convert_expr
+from cvxpy.reductions.solvers.nlp_solvers.diff_engine.converters import convert_expr
 from cvxpy.reductions.solvers.nlp_solvers.diff_engine.helpers import (
     build_param_dict,
     build_var_dict,
@@ -32,7 +32,6 @@ from cvxpy.reductions.solvers.nlp_solvers.diff_engine.registry import ATOM_CONVE
 __all__ = [
     "C_problem",
     "ATOM_CONVERTERS",
-    "build_capsule",
     "build_var_dict",
     "build_param_dict",
     "convert_expr",
