+++ cvxpy/reductions/solvers/nlp_solvers/diff_engine/c_problem.py
@@ -17,7 +17,11 @@ from sparsediffpy import _sparsediffengine as _diffengine
 
 import cvxpy as cp
 from cvxpy.reductions.inverse_data import InverseData
-from cvxpy.reductions.solvers.nlp_solvers.diff_engine.converters import build_capsule
+from cvxpy.reductions.solvers.nlp_solvers.diff_engine.converters import convert_expr
+from cvxpy.reductions.solvers.nlp_solvers.diff_engine.helpers import (
+    build_param_dict,
+    build_var_dict,
+)
 
 
 class C_problem:
@@ -31,10 +35,25 @@ class C_problem:
             verbose: print solver output
         """
         inverse_data = InverseData(cvxpy_problem)
-        constraint_exprs = [arg for c in cvxpy_problem.constraints for arg in c.args]
-        self._capsule, _, _ = build_capsule(
-            cvxpy_problem.objective.expr, constraint_exprs, inverse_data,
-            params=cvxpy_problem.parameters(), verbose=verbose)
+        var_dict, n_vars = build_var_dict(inverse_data)
+        param_dict = build_param_dict(cvxpy_problem, inverse_data)
+
+        c_obj = convert_expr(cvxpy_problem.objective.expr,
+                             var_dict, n_vars, param_dict)
+        c_constraints = [convert_expr(c.expr, var_dict, n_vars, param_dict)
+                         for c in cvxpy_problem.constraints]
+        self._capsule = _diffengine.make_problem(
+            c_obj, c_constraints, verbose)
+
+        if param_dict:
+            _diffengine.problem_register_params(
+                self._capsule, list(param_dict.values()))
+            # Set initial parameter values
+            theta = np.concatenate([
+                np.asarray(p.value, dtype=np.float64).flatten(order='F')
+                for p in cvxpy_problem.parameters()
+            ])
+            _diffengine.problem_update_params(self._capsule, theta)
 
     def update_params(self, theta: np.ndarray) -> None:
         """Update parameter values in the C DAG.
